requirements.txt
openai==1.54.4
python-dotenv==1.0.1
flask==3.1.0
pymupdf==1.24.14
requests==2.32.3 
boto==1.35.68